+++ modules/sys/macro.js
@@ -1,7 +1,4 @@
-import { MODULE_ID, MACRO_VERS } from './constants.js';
-import { HMDialogMgr } from '../mgr/dialogmgr.js';
-import { HMChatMgr } from '../mgr/chatmgr.js';
-import { HMRollMgr } from '../mgr/rollmgr.js';
+import { MODULE_ID } from './constants.js';
 
 async function createItemMacro(data, slot) {
     if (!game || !game.macros) return;
@@ -36,38 +33,6 @@ async function createItemMacro(data, slot) {
 }
 
 export class HMMacro extends Macro {
-    isObsolete(d_vers, d_mid) {
-        const vers = parseInt(DOMPurify.sanitize(d_vers), 10);
-        const mid  = DOMPurify.sanitize(d_mid);
-        if (MACRO_VERS?.[mid] === vers) return false;
-
-        const msg = `Your macro, <b> ${this.name}</b>, is obsolete. `
-                  + `Please obtain a fresh copy from the system compendium.`;
-        ui.notifications.error(msg);
-        return true;
-    }
-
-    get dialogmgr() { return new HMDialogMgr }
-    get chatmgr()   { return new HMChatMgr }
-    get rollmgr()   { return new HMRollMgr }
-
-    _hm_getActor({allActors=false}={}) {
-        const tokens = canvas.tokens.controlled.filter((a) => a.actor);
-
-        let actors = [];
-        for (let i = 0; i < tokens.length; i++) {
-            if (tokens[i]?.actor) actors.push(tokens[i]?.actor);
-        }
-        if (!actors.length) return null;
-        if (allActors) return actors;
-
-        const actor=actors[0];
-        if (actors.length > 1) {
-            ui.notifications.warn(`${game.i18n.localize('HM.dialog.warnMulti')} ${actor.name}</b>.`);
-        }
-        return actor;
-    }
-
     static hotbarDrop(_bar, data, slot) {
         if (data.type === 'Item') {
             createItemMacro(data, slot);