+++ CHANGELOG.md
@@ -4,7 +4,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
-- Weapon cards have a button to indicate which weapon displays reach.
+- Weapon cards have a button to indicate/select which weapon displays reach.
 - Weapon cards display ranged weapon range breakdown.
 ### Changed
 - Reach/threat ranges highlight on mouseover.