+++ CHANGELOG.md
@@ -4,6 +4,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Changed
+- Compendium items.
 - Re-styled character class sheet.
 - Character sheet style improvements.
 ### Fixed