+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 - Scrollable sections of the character sheet no longer reset.
 - Possible crash due to token redraws in rare circumstances.
+- Revised Inflict Wound macro.
 
 ## [0.2.27] - 2022-11-01
 ### Changed