+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [0.2.28] - 2022-11-03
 ### Added
 - Button to toggle weapon equip state from Combat tab.
 ### Changed